updating kde process
Now users can input a year/month and it will rerun the KDE estimate for all year/months greater than inputted value. If a user wants to add a KDE to the layer after new eviction data is added to the server, once the server data has been updated, the user can simply enter the year/month to add a KDE raster file to the web-map.